+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **Added**: Confiration dialog for appling identifier info
 - **Added**: Edit and add for feeds
 - **Changed**: Viewer version from 0.4.21 to 0.4.24
+- **Changed**: Year to published_at
 - **Changed**: In `AdminAddEntry` page in step one (identifiers) 2 inputs to 1
 - **Changed**: Entries loading
 - **Changed**: Modal